submit button
changed button type to submit and set the onsubmit attribute of the form to return false to stop refreshing page when button is clicked. it's not *quite* perfect because the error/validation message appears automatically after the alert is created. next will try playing with event listeners to see if i can solve that.