button click event listener
added id to button and trying to experiment with using the click event as our listener that we trigger functions off of. so far i can get it to display to console but having trouble getting it to recognize the task input field and check it's validity as part of the function.